+++ src/chemicalUnits/SupraMolecule.cpp
@@ -222,7 +222,7 @@ void SupraMolecule::removeMolecule(Molecule molecule){
 
 double SupraMolecule::getSupraMolecularMass(){
     double supraMolecularMass = 0;
-    for (int i = 0; i < this->supraMolecule.size(); i++){
+    for (int i = 0; i < (int) this->supraMolecule.size(); i++){
         supraMolecularMass = supraMolecularMass + supraMolecule[i].getMolecularMass();
     };
     return supraMolecularMass;
